# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Holivita AI API",
        "description": "API для онбординг-чата с AI-ассистентом, суммаризации и голосового ввода (OpenAI Whisper). Авторизация отключена для MVP — user_id передаётся вручную. Модули: Onboarding (синхронный и асинхронный чат), Summaries (суммаризации сессий), Voice (транскрипция аудио). Формат ответов: успех — {success: true, data: {...}}, ошибка — {success: false, error: \"...\"}.",
        "contact": {
            "name": "Holivita AI Team",
            "email": "support@holivita.ai"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/voice/transcribe": {
            "post": {
                "tags": [
                    "Voice"
                ],
                "summary": "Синхронная транскрипция аудио в текст",
                "description": "Отправляет аудиофайл на транскрипцию через OpenAI Whisper API и возвращает распознанный текст синхронно. Поддерживаемые форматы: webm, wav, mp3, mp4, ogg, flac. Максимальный размер файла — 25 МБ. Если речь не распознана — возвращает пустой текст с сообщением.",
                "operationId": "voiceTranscribe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "audio",
                                    "user_id"
                                ],
                                "properties": {
                                    "audio": {
                                        "description": "Аудиофайл (webm, wav, mp3, mp4, ogg, flac). Макс. 25 МБ.",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_id": {
                                        "description": "ID пользователя",
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 1
                                    },
                                    "language": {
                                        "description": "Код языка ISO 639-1 (по умолчанию: ru)",
                                        "type": "string",
                                        "maxLength": 2,
                                        "example": "ru"
                                    },
                                    "session_id": {
                                        "description": "ID сессии онбординга (опционально, для привязки к сессии)",
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Транскрипция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "text": {
                                                    "description": "Распознанный текст (пустая строка, если речь не распознана)",
                                                    "type": "string",
                                                    "example": "Меня зовут Иван, мне 35 лет"
                                                },
                                                "language": {
                                                    "description": "Определённый язык аудио",
                                                    "type": "string",
                                                    "example": "ru"
                                                },
                                                "confidence": {
                                                    "description": "Уровень уверенности распознавания (0–1)",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0.95,
                                                    "nullable": true
                                                },
                                                "duration": {
                                                    "description": "Длительность аудио в секундах",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 3.5,
                                                    "nullable": true
                                                },
                                                "provider": {
                                                    "description": "Провайдер транскрипции",
                                                    "type": "string",
                                                    "example": "openai"
                                                },
                                                "message": {
                                                    "description": "Сообщение (только если текст пустой)",
                                                    "type": "string",
                                                    "example": "Не удалось распознать речь. Попробуйте ещё раз."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации (неверный формат, превышен размер и т.д.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Аудио файл обязателен."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Сервис транскрипции недоступен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Сервис транскрипции временно недоступен."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/voice/async/transcribe": {
            "post": {
                "tags": [
                    "Voice"
                ],
                "summary": "Асинхронная транскрипция аудио",
                "description": "Отправляет аудиофайл в очередь для асинхронной транскрипции. Аудио сохраняется на диск, и задача ставится в очередь. Ответ возвращается немедленно с ID транскрипции и статусом pending. Для получения результата используйте GET /voice/async/status. Если предыдущая транскрипция ещё обрабатывается — возвращается ошибка 409.",
                "operationId": "voiceAsyncTranscribe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "audio",
                                    "user_id"
                                ],
                                "properties": {
                                    "audio": {
                                        "description": "Аудиофайл (webm, wav, mp3, mp4, ogg, flac). Макс. 25 МБ.",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_id": {
                                        "description": "ID пользователя",
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 1
                                    },
                                    "language": {
                                        "description": "Код языка ISO 639-1 (по умолчанию: ru)",
                                        "type": "string",
                                        "maxLength": 2,
                                        "example": "ru"
                                    },
                                    "session_id": {
                                        "description": "ID сессии онбординга (опционально)",
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Аудио принято в обработку",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "transcription_id": {
                                                    "description": "ID транскрипции для отслеживания статуса",
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "660e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "pending"
                                                    ],
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Предыдущая транскрипция ещё обрабатывается",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Предыдущая транскрипция ещё обрабатывается."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Аудио файл обязателен."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}